+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/ArtifactResolver.java
@@ -19,6 +19,7 @@ import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toSet;
 
 import java.io.File;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
@@ -32,7 +33,7 @@ import org.eclipse.aether.artifact.Artifact;
 import org.eclipse.aether.artifact.DefaultArtifact;
 import org.eclipse.aether.collection.CollectRequest;
 import org.eclipse.aether.graph.Dependency;
-import org.eclipse.aether.graph.DependencyFilter;
+import org.eclipse.aether.graph.Exclusion;
 import org.eclipse.aether.repository.RemoteRepository;
 import org.eclipse.aether.resolution.ArtifactResult;
 import org.eclipse.aether.resolution.DependencyRequest;
@@ -41,8 +42,7 @@ import org.eclipse.aether.resolution.DependencyResult;
 
 public class ArtifactResolver {
 
-	private static final DependencyFilter ACCEPT_ALL = (node, parents) -> true;
-	private static final DependencyFilter FILTER_TRANSITIVES = (node, parents) -> parents.size() <= 1;
+	private final static Exclusion EXCLUDE_ALL_TRANSITIVES = new Exclusion("*", "*", "*", "*");
 
 	private final RepositorySystem repositorySystem;
 	private final RepositorySystemSession session;
@@ -68,18 +68,16 @@ public class ArtifactResolver {
 	 * of the specified coordinates and optionally their transitive dependencies.
 	 */
 	public Set<File> resolve(boolean withTransitives, Collection<String> mavenCoordinates) {
+		Collection<Exclusion> excludeTransitive = new ArrayList<Exclusion>(1);
+		if (!withTransitives) {
+			excludeTransitive.add(EXCLUDE_ALL_TRANSITIVES);
+		}
 		List<Dependency> dependencies = mavenCoordinates.stream()
 				.map(coordinateString -> new DefaultArtifact(coordinateString))
-				.map(artifact -> new Dependency(artifact, null))
+				.map(artifact -> new Dependency(artifact, null, null, excludeTransitive))
 				.collect(toList());
 		CollectRequest collectRequest = new CollectRequest(dependencies, null, repositories);
-		DependencyRequest dependencyRequest;
-		if (withTransitives) {
-			dependencyRequest = new DependencyRequest(collectRequest, ACCEPT_ALL);
-		} else {
-			dependencyRequest = new DependencyRequest(collectRequest, FILTER_TRANSITIVES);
-		}
-
+		DependencyRequest dependencyRequest = new DependencyRequest(collectRequest, null);
 		DependencyResult dependencyResult = resolveDependencies(dependencyRequest);
 
 		return dependencyResult.getArtifactResults()