+++ plugin-maven/src/test/java/com/diffplug/spotless/maven/MavenIntegrationTest.java
@@ -106,6 +106,8 @@ public class MavenIntegrationTest extends ResourceHarness {
 		writePom(groupWithSteps("kotlin", steps));
 	}
 
+	/** The XML extension is discontinued. */
+	@Deprecated
 	protected void writePomWithXmlSteps(String... steps) throws IOException {
 		writePom(groupWithSteps("xml", steps));
 	}
@@ -114,6 +116,8 @@ public class MavenIntegrationTest extends ResourceHarness {
 		writePom(groupWithSteps("cpp", steps));
 	}
 
+	/** The CSS extension is discontinued. */
+	@Deprecated
 	protected void writePomWithCssSteps(String... steps) throws IOException {
 		writePom(groupWithSteps("css", steps));
 	}